Combine fixtures (#431)
Move fixtures to conftest.py so they are shared.
