+++ chromadb/test/property/test_collections.py
@@ -1,29 +1,16 @@
 import pytest
 import logging
-from hypothesis import given, assume, settings
 import hypothesis.strategies as st
-from typing import List
-import chromadb
-from chromadb.api import API
-from chromadb.api.models.Collection import Collection
-from chromadb.test.configurations import configurations
 import chromadb.test.property.strategies as strategies
 from hypothesis.stateful import (
     Bundle,
     RuleBasedStateMachine,
     rule,
     initialize,
-    precondition,
     multiple,
     consumes,
     run_state_machine_as_test,
 )
-
-
-@pytest.fixture(scope="module", params=configurations())
-def api(request):
-    configuration = request.param
-    return chromadb.Client(configuration)
 
 
 class CollectionStateMachine(RuleBasedStateMachine):
@@ -120,7 +107,6 @@ class CollectionStateMachine(RuleBasedStateMachine):
         return coll
 
 
-# TODO: takes 7-8 minutes to run, figure out how to make faster. It shouldn't take that long, it's only 3-5000 database operations and DuckDB is faster than that
 def test_collections(caplog, api):
     caplog.set_level(logging.ERROR)
     run_state_machine_as_test(lambda: CollectionStateMachine(api))
